Increase Microsoft logo size to match other provider icons
